requirements.txt
Django>=4.2
djangorestframework
drf-yasg
gunicorn
django-storages
boto3
psycopg2-binary
django-redis
channels
channels-redis
pytest
pytest-django
pytest-asyncio
daphne